+++ src/assets/jss/material-dashboard-react/components/customInputStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // CustomInput styles
-// #############################
-
 import {
   primaryColor,
   dangerColor,