+++ src/assets/jss/material-dashboard-react/components/footerStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // Footer styles
-// #############################
-
 import { defaultFont, container, primaryColor } from "assets/jss/material-dashboard-react.jsx";
 
 const footerStyle = {