+++ src/assets/jss/material-dashboard-react/components/headerLinksStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // HeaderLinks styles
-// #############################
-
 import {
   defaultFont,
   dangerColor,