+++ src/assets/jss/material-dashboard-react/components/sidebarStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // Sidebar styles
-// #############################
-
 import {
   drawerWidth,
   transition,