+++ src/assets/jss/material-dashboard-react/components/snackbarContentStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // SnackbarContent styles
-// #############################
-
 import {
   defaultFont,
   primaryBoxShadow,