+++ src/assets/jss/material-dashboard-react/components/tableStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // Table styles
-// #############################
-
 import {
   warningColor,
   primaryColor,