+++ src/assets/jss/material-dashboard-react/components/typographyStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // Typography styles
-// #############################
-
 import {
   defaultFont,
   primaryColor,