+++ src/assets/jss/material-dashboard-react/layouts/dashboardStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // App styles
-// #############################
-
 import { drawerWidth, transition, container } from "assets/jss/material-dashboard-react.jsx";
 
 const appStyle = theme => ({