+++ src/assets/jss/material-dashboard-react/views/dashboardStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // Dashboard styles
-// #############################
-
 import { successColor } from "assets/jss/material-dashboard-react.jsx";
 
 const dashboardStyle = {