+++ src/components/CustomButtons/Button.jsx
@@ -5,7 +5,7 @@ import PropTypes from "prop-types";
 import withStyles from "@material-ui/core/styles/withStyles";
 import Button from "@material-ui/core/Button";
 // core components
-import buttonStyle from "assets/jss/material-dashboard-react/buttonStyle";
+import buttonStyle from "assets/jss/material-dashboard-react/components/buttonStyle";
 
 function RegularButton({ ...props }) {
   const {