+++ src/components/Footer/Footer.jsx
@@ -5,7 +5,7 @@ import withStyles from "@material-ui/core/styles/withStyles";
 import ListItem from "@material-ui/core/ListItem";
 import List from "@material-ui/core/List";
 // core components
-import footerStyle from "assets/jss/material-dashboard-react/footerStyle";
+import footerStyle from "assets/jss/material-dashboard-react/components/footerStyle";
 
 function Footer({ ...props }) {
   const { classes } = props;