+++ src/components/Snackbar/SnackbarContent.jsx
@@ -8,7 +8,7 @@ import IconButton from "@material-ui/core/IconButton";
 // @material-ui/icons
 import Close from "@material-ui/icons/Close";
 // core components
-import snackbarContentStyle from "assets/jss/material-dashboard-react/snackbarContentStyle.jsx";
+import snackbarContentStyle from "assets/jss/material-dashboard-react/components/snackbarContentStyle.jsx";
 
 function SnackbarContent({ ...props }) {
   const { classes, message, color, close, icon } = props;