+++ src/components/Tasks/Tasks.jsx
@@ -14,7 +14,7 @@ import Edit from "@material-ui/icons/Edit";
 import Close from "@material-ui/icons/Close";
 import Check from "@material-ui/icons/Check";
 // core components
-import tasksStyle from "assets/jss/material-dashboard-react/tasksStyle.jsx";
+import tasksStyle from "assets/jss/material-dashboard-react/components/tasksStyle.jsx";
 
 class Tasks extends React.Component {
   state = {