+++ src/components/Typography/Warning.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 // @material-ui/core components
 import withStyles from "@material-ui/core/styles/withStyles";
 // core components
-import typographyStyle from "assets/jss/material-dashboard-react/typographyStyle.jsx";
+import typographyStyle from "assets/jss/material-dashboard-react/components/typographyStyle.jsx";
 
 function Warning({ ...props }) {
   const { classes, children } = props;