+++ src/layouts/Dashboard/Dashboard.jsx
@@ -13,7 +13,7 @@ import Sidebar from "components/Sidebar/Sidebar.jsx";
 
 import dashboardRoutes from "routes/dashboard.jsx";
 
-import appStyle from "assets/jss/material-dashboard-react/appStyle.jsx";
+import dashboardStyle from "assets/jss/material-dashboard-react/layouts/dashboardStyle.jsx";
 
 import image from "assets/img/sidebar-2.jpg";
 import logo from "assets/img/reactlogo.png";
@@ -85,4 +85,4 @@ App.propTypes = {
   classes: PropTypes.object.isRequired
 };
 
-export default withStyles(appStyle)(App);
+export default withStyles(dashboardStyle)(App);