+++ src/views/Dashboard/Dashboard.jsx
@@ -30,7 +30,7 @@ import {
   completedTasksChart
 } from "variables/charts";
 
-import dashboardStyle from "assets/jss/material-dashboard-react/dashboardStyle";
+import dashboardStyle from "assets/jss/material-dashboard-react/views/dashboardStyle.jsx";
 
 class Dashboard extends React.Component {
   state = {